# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WorkShield API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/reviews": {
      "post": {
        "tags": [
          "reviews"
        ],
        "summary": "Start Review",
        "description": "Cookie 소유 세션의 검토를 fingerprint 기준으로 멱등 접수한다.",
        "operationId": "start_review_api_v1_reviews_post",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCreateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ReviewCreateResponse_"
                }
              }
            }
          },
          "404": {
            "description": "리소스가 없거나 현재 익명 세션에서 접근할 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "현재 상태 또는 멱등 키 충돌",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "410": {
            "description": "익명 세션 만료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "요청 값 검증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "외부 서비스 사용 불가",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "외부 서비스 시간 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews/{review_id}/results": {
      "get": {
        "tags": [
          "reviews"
        ],
        "summary": "Get Results",
        "operationId": "get_results_api_v1_reviews__review_id__results_get",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Review Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ReviewResponse_"
                }
              }
            }
          },
          "404": {
            "description": "리소스가 없거나 현재 익명 세션에서 접근할 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "현재 상태 또는 멱등 키 충돌",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "410": {
            "description": "익명 세션 만료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "요청 값 검증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "외부 서비스 사용 불가",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "외부 서비스 시간 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews/{review_id}/grounding": {
      "get": {
        "tags": [
          "reviews"
        ],
        "summary": "Get Grounding",
        "description": "현재 검토 결과의 category에 해당하는 법령 참고자료를 조회한다.",
        "operationId": "get_grounding_api_v1_reviews__review_id__grounding_get",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Review Id"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_GroundingResponse_"
                }
              }
            }
          },
          "404": {
            "description": "리소스가 없거나 현재 익명 세션에서 접근할 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "현재 상태 또는 멱등 키 충돌",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "410": {
            "description": "익명 세션 만료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "요청 값 검증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "외부 서비스 사용 불가",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "외부 서비스 시간 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews/{review_id}/chat/messages": {
      "post": {
        "tags": [
          "reviews"
        ],
        "summary": "Chat Message",
        "description": "대화 본문을 별도 영구 이력으로 남기지 않는 검토 범위 질의응답.",
        "operationId": "chat_message_api_v1_reviews__review_id__chat_messages_post",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Review Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ChatResponse_"
                }
              }
            }
          },
          "404": {
            "description": "리소스가 없거나 현재 익명 세션에서 접근할 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "현재 상태 또는 멱등 키 충돌",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "410": {
            "description": "익명 세션 만료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "요청 값 검증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "외부 서비스 사용 불가",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "외부 서비스 시간 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews/{review_id}/suggestions": {
      "post": {
        "tags": [
          "reviews"
        ],
        "summary": "Create Suggestion",
        "description": "검증된 단일 사용자·표준조항과 grounding으로 협의 초안을 생성한다.",
        "operationId": "create_suggestion_api_v1_reviews__review_id__suggestions_post",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Review Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuggestionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SuggestionResponse_"
                }
              }
            }
          },
          "404": {
            "description": "리소스가 없거나 현재 익명 세션에서 접근할 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "현재 상태 또는 멱등 키 충돌",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "410": {
            "description": "익명 세션 만료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "요청 값 검증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "외부 서비스 사용 불가",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "외부 서비스 시간 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews/{review_id}/events": {
      "get": {
        "tags": [
          "reviews"
        ],
        "summary": "Review Events",
        "description": "저장된 MCP progress를 단조 sequence의 SSE로 전달한다.",
        "operationId": "review_events_api_v1_reviews__review_id__events_get",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Review Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "리소스가 없거나 현재 익명 세션에서 접근할 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "현재 상태 또는 멱등 키 충돌",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "410": {
            "description": "익명 세션 만료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "요청 값 검증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "외부 서비스 사용 불가",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "외부 서비스 시간 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "type": "object",
        "required": [
          "error",
          "meta"
        ],
        "title": "ApiErrorResponse",
        "description": "JSON API의 공통 오류 응답."
      },
...
      "ApiResponse_ChatResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ChatResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[ChatResponse]"
      },
      "ApiResponse_GroundingResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GroundingResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[GroundingResponse]"
      },
...
      "ApiResponse_ReviewCreateResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ReviewCreateResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[ReviewCreateResponse]"
      },
      "ApiResponse_ReviewResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ReviewResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[ReviewResponse]"
      },
...
      "ChatRequest": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Message"
          },
          "focus_clause_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Focus Clause Id"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/ChatHistoryMessage"
            },
            "type": "array",
            "maxItems": 10,
            "title": "History"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatRequest"
      },
...
      "ReviewCreateRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Session Id"
          }
        },
        "type": "object",
        "required": [
          "session_id"
        ],
        "title": "ReviewCreateRequest",
        "description": "검토 시작 요청."
      },
...
      "SuggestionRequest": {
        "properties": {
          "user_clause_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "User Clause Id"
          },
          "purpose": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Purpose"
          },
          "inputs": {
            "additionalProperties": true,
            "type": "object",
            "title": "Inputs"
          }
        },
        "type": "object",
        "required": [
          "user_clause_id",
          "purpose"
        ],
        "title": "SuggestionRequest"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "workshield_session"
      }
    }
  }
}